MR-75: Fix delete function and create button for Add member
